Rename: contant -> *.constant
