+++ webapp/src/api/upload.api.js
@@ -1,11 +1,23 @@
 import privateApiInstance from './instance/privateApiInstance';
 
 const uploadApi = {
-  IMAGE_UPLOAD(data) {
+  uploadImage(data) {
+    return privateApiInstance({
+      url: '/upload',
+      headers: {
+        'Content-Type': 'multipart/form-data',
+      },
+      method: 'post',
+      data,
+    });
+  },
+  deleteImage(imgUrl) {
     return privateApiInstance({
       url: '/upload',
       method: 'post',
-      data,
+      params: {
+        imgUrl,
+      },
     });
   },
 };
