+++ webapp/src/components/GlobalNavigation/UserInfoDropdown.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { useNavigate } from 'react-router-dom';
 import Dropdown from 'components/Common/Dropdown';
 import { deleteUserInfo } from 'service/auth';
-import { MY_POST, PROFILE, MY_LIST } from 'constant/route';
+import { MY_POST, PROFILE, MY_LIST } from 'constant/route.constant';
 import authApi from 'api/auth.api';
 import * as S from './style';
