+++ webapp/src/mocks/commentHandler.mock/index.js
@@ -1,4 +1,4 @@
-import { API, ROOT_API_URL } from 'constant/api';
+import { API, ROOT_API_URL } from 'constant/api.constant';
 import { getResonseWithData, randomResponse } from 'mocks/mockUtils';
 import { rest } from 'msw';
 import { teamComments } from './teamComments';
